This solves the Blinder uploading error.

+++ grails-app/controllers/com/deerwalk/testdatagenerator/BlindingDataController.groovy
@@ -33,19 +33,30 @@ class BlindingDataController {
 
     def upload() {
         //this delete all previous files in the blinder-->> input,jsonFormat and output dir.
-        new File("src/main/resources/blinder/input").eachFileMatch(~/.*.csv/) { file ->
-            file.delete()
+        def csvFile =  new File("src/main/resources/blinder/input")
+        if(csvFile.list()){
+            csvFile.eachFileMatch(~/.*.csv/) { file ->
+                file.delete()
+            }
         }
-        new File("src/main/resources/blinder/jsonFormat").eachFileMatch(~/.*.json/) { file ->
-            file.delete()
+
+        def jsonFile = new File("src/main/resources/blinder/jsonFormat")
+        if(jsonFile.list()){
+            jsonFile.eachFileMatch(~/.*.json/) { file ->
+                file.delete()
+            }
         }
-        new File("src/main/resources/blinder/output").eachFileMatch(~/.*.csv/) { file ->
-            file.delete()
+
+        def outputFile = new File("src/main/resources/blinder/output")
+        if(outputFile.list()){
+            outputFile.eachFileMatch(~/.*.csv/) { file ->
+                file.delete()
+            }
         }
+
 
         def f = request.getFile('fileUpload')
         if(!f.empty) {
-            flash.message = 'Your file has been uploaded'
             new File( grailsApplication.config.images.location.toString() ).mkdirs()
             def uploadFileName = f.getOriginalFilename()
             def locationUpload = "src/main/resources/blinder/input/"+ uploadFileName
@@ -53,12 +64,14 @@ class BlindingDataController {
 //            f.transferTo( new File( "src/Upload/" + f.getOriginalFilename() ) )
 
             def onlyFileName = StringUtils.substringBefore(uploadFileName, ".")
+
+
             //update the logic
             FindTypeOfData.FindTheType(locationUpload, "src/main/resources/blinder/jsonFormat/"+onlyFileName+".json")
-
             //to create new output file.
             def file = new File("src/main/resources/blinder/output/blinded_"+onlyFileName+".csv")
             file.createNewFile()
+            flash.message = 'Your file has been uploaded'
         }
         else {
             flash.message = 'file cannot be empty'
